requirements.txt
streamlit==1.37.0

langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33

langchain-groq==0.1.5
langchain-huggingface==0.0.3
langchain-chroma==0.1.2

chromadb==0.5.5

sentence-transformers==3.0.1

python-dotenv==1.0.1

pypdf==4.3.1